Implemented LoadReport() method
With this method you can now load an already generated report
and parse it via defined XSLT files more easily.

The root node of the loaded report must match the tag name given
in the XMLOut constructor.  LoadReport can also be set to
validate the version tag against what's set as version in the
constructor as well.