Reworked the XML-RPC server part to work with mod_python
The apache-htaccess is the needed .htaccess file for the
Apache web server.  It should be located in the directory where
also rteval_xmlrpc.py and xmlrpc_API1.py should reside.  This
directory must be available from a HTTP request.  Please make
sure that this directory are allowed to use .htaccess files by
the Apache config.

In addition, the rtevald.py was adjusted to work as a local
test server, to avoid to configure a Apache + mod_python locally

Also added some copyright information into the touched files
as well