Implemented Python part of XML -> SQL parser
This module will prepare a dictionary which can be
used by a Database object for quering a database.

   class XMLSQLparser(XSLT, XMLdata)
	def GetSQLdata(table, rterid=None, syskey=None)

The XSLT is a filename to a xmlparser.xsl file, which defines the
mapping of XML fields into SQL tables and fields.  The XMLdata parameter
is a filename to the XML data being processed.  Future version will support
both filename and xmlDoc.

The XMLSQLparser::GetSQLdata(...) function will parse the XML data into
a Python dictionary containing all needed field names, values and table name
which is useable by a Database object.  The table parameter is required, and
a matching table name must be defined in the XSLT template.  The rterid and
syskey paramters are optional, but some table definitions might need them.  For
more info, look into the xmlparser.xsl file