Improved robustness of database connections and thread control
Both the process_submission_queue() and parserthread() functions
now make use of db_ping() to make sure they have an alive database
connection.  And if not, trigger the needed notifications.

If the process_submission_queue() looses the database connection
completely, it will shutdown rteval_parserd immediately.

If the parserthread()s looes the database connection, it will
reduce the activethreads counter.  And if this counter reaches 0,
it will signal the main thread which initiates a complete shutdown
of rteval_parserd.

Btw. db_ping() will indeed try to restore the database connection
before presuming the database is unavailable.