Added new config parameter: max_report_size
To avoid potential DoS attacks by sending big XML files, the max_report_size
rejects reports which is bigger than this value.  The default value is 2MB if
this value is not configured.

A new status code was introduced, STAT_FTOOBIG (12).  The submissionqueue
record will be updated with this value on records which fails this check.

In addition minor SIGSEGV failure was fixed when rteval_parserd shutsdown
before worker threads have really started.