Implemented SQL schema version handling on the database layer
When a database connection is established, the rteval_info table is queried
for a key/value pair named 'sql_schema_ver' and parsed.  This attribute defines
which version the rteval database schema is in use.

When processing XML documents with <sqldata/>, a checks is done to make sure
there are no attempts on inserting data to an incompatible rteval database schema.

If the rteval_info table or the sql_schema_ver value is not found, it will default
to version 1.0 (the parsed int value will be 100)