Avoid unexpected exit if the main thread starts process_submission_queue() too early
If the process_submission_queue() function in the main thread is started before
the parsethread() based threads, the number of active threads will be 0 and
process_submission_queue() will immediately send a "shutdown" message and exit.

The quick solution is to sleep 3 seconds before starting the
process_submission_queue() job.  This way at least a few of the parsethread()
threads should have had time to start up and setle.