Added unit-tests
Added a generic unit-test framework and rewritten the self-test parts in
some rteval modules where this was appropriate.

The unit-test contains a list of modules to load and test. It will run a
function named unit_test() in the imported modules.  It is expected that
the unit_test() function will return 0 on success, otherwise it is logged
as an error.