Added mod_wsgi handler
This is an alternative Python handler for the XML-RPC server, making
use of the WSGI platform instead.

If the rteval_xmlrpc.wsgi script is started manually, using python
it will act as a standalone server, listening to localhost:65432
as default.

This script can also be loaded by, f.ex., the mod_wsgi module to
Apache.  A template for Apache config is also added.

Signed-off-by: David Sommerseth <davids@redhat.com>