server/parser: Make the parsing of measurement data more flexible
The parsing of measurement data was hard coded to be very cyclictest
centric.  With this patch, the code paths used by cyclictest parsing
is made more generic and which measurement data to parse is now
declared in the configuration file.

The default measurement data tables are set to be cyclic_statistics,
cyclic_histogram and cyclic_rawdata.  These tables can be overridden
by setting the measurement_tables config variable in the
xmlrpc_parser section.

With this change, it is now possible to extend the rteval summary.xml
measurement parsing without modifying the C code at all.  Only the
database schema and the xmlparser.xsl files needs to be updated, and
then add this new table to configuration file.

Signed-off-by: David Sommerseth <davids@redhat.com>