Changes the getcmdpath method to use only python calls to find the paths
In my previous patches, I introduced a getcmdpath method which intended
to replace the direct calls to binary from clear paths, using the which
binary to locate the place of the searched one. This behavior was in
fact not the finest as the method itself used the full path to the which
binary. In this patch, I corrected the getcmdpath command to use only
python calls (os.path.isfile and os.access) to identify in the PATH
environment variable the place where the binary is.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>
Signed-off-by: Clark Williams <williams@redhat.com>