Revamped the onlyload flag handling
Don't call RtEval::measure() when --onlyload are given.  This
simplifies the measure() method and makes its purpose cleaner.

This change also requires --logging to be used to create a report
directory when being in --onlyload mode.  But it will not create
a tarball in this case or any other types of reports.

Signed-off-by: David Sommerseth <davids@redhat.com>