Rework module imports once more - save module references again
The measurement modules needs to provide some information before
they can be instantiated as objects.  So rework the ModuleContainer()
class to save references to the modules which can be reused again and
which helps avoiding trying to re-import already imported modules.

Added a ModuleContainer::ModuleInfo() method, which calls the imported
module's ModuleInfo() function and passes the result from the function
back to the caller.

Renamed the ModuleContainer::Import() and RtEvalModule::_Import()
methods to InstantiateModule(), to indicate clearer what the method
does.

Signed-off-by: David Sommerseth <davids@redhat.com>