Added a rtevalModulePrototype() class
This class defines the needed methods for both rteval load and measurement
modules.  It also implements the needed run logic, so the modules just
need to implement those code pieces unique for the module.

The needed methods to implement are:

_WorkloadSetup()   - Prepares for building/compiling additional code
_WorkloadBuild()   - Does the building/compiling.  This method needs to
                     call self._setReady() when it is done.
_WorkloadPrepare() - Final preparations before the workload is kicked off
_WorkloadTask()    - Is called in a loop until the rteval duration has
                     expired.  This should kick off the core workload.
_WorkloadAlive()   - Implements the needed check to see that the workload
                     is running
_WorkloadCleanup() - Called after rteval have stopped the modules' workload
                     and should save the needed information for the final
                     report back to rteval
MakeReport()       - This method generates the report and returns it as a
                     libxml2.xmlNode object.

This patch also adopts the RtEvalModules() class to depend on the API of
rtevalModulePrototype().

Signed-off-by: David Sommerseth <davids@redhat.com>