Don't raise an exception if measurement threads stops running
There might be reasons why measurement threads don't complete the
set duration time.  If this is an error, that should be handled in
the measurement module and not globally.

If one or more measurement threads have stopped before the duration time
has expired, rather do a controlled exit - generating a report.

Signed-off-by: David Sommerseth <davids@redhat.com>