Split up rteval.py into a pure measurement module and the command line tool
This cleans up rteval further by making the RtEval class a pure measurement
feature which receives all the needed setup and controls loads and measurment
modules, in addition to the final reporting.

The command line and config parsing is kept inside the command line tool, where
it will set up everything before passing it all to the RtEval calss.

Signed-off-by: David Sommerseth <davids@redhat.com>