Added infrastructure to add command line arguments from modules
All rteval modules must now provide a ModuleParameters() function
which returns a dictionary containing the option setup it expects.

Each module will get a separate option group, and all its options
will be prefixed with the module name.  These option variables will
be stored as well, prefixed with the module name.  So if the a module
'dummy' configures an option 'value1', it will be stored as
'dummy_value1' in the option container.  To set this option, you
will need to use --dummy-value1 from the command line.

Signed-off-by: David Sommerseth <davids@redhat.com>