Changes the default_config to allow running rteval from git/extracted directory without options
This commit introduce a new method default_config_search to search, for
each file needed in the configuration, the better default place. This place
can be the rteval file directory, the rteval file directory appended with the
'rteval' directory, the ~/.rteval directory of the user, the /etc/rteval
directory, or the /usr/share/rteval directory.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>
Signed-off-by: David Sommerseth <davids@redhat.com>