Moves the update of the log verbosity to use it properly during modules instantiation
Previously, we used a default Log.INFO verbose level before parsing command line
arguments, as modules load is necessary to know all the available arguments. In this
patch, we added an small command line analysis for the verbose and debug arguments to
set properly the log verbosity before loading modules. Doing so, we are able to have
the wanted verbosity level before starting the load of modules, and not to have
unwanted messages during execution.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>
Signed-off-by: David Sommerseth <davids@redhat.com>