Put the rteval version into a separate importable file
This is to avoid having massive build requirements just to retrieve
the rteval version.

The version string is also still available via the rteval module as
well, to avoid other external implementations to fail with this change.
But that variable is basically just a "re-export" of the value found
in rteval/version.py

Signed-off-by: David Sommerseth <davids@redhat.com>