Reworked the run logic for modules - allow them to flag they won't run
If a module sets the self._donotrun flag to True, the module logic will
take care of not running this module at all.

This also sorts out some issues when hackbench is not run, where some of
the log file descriptors are not found when cleaning up in the end.

Signed-off-by: David Sommerseth <davids@redhat.com>