Add a 30 seconds sleep before unleashing the measurement threads
This is to help the system to settle down before kicking off the
realtime measurement workload.  Otherwise some abnormal latency
spikes might be logged, which would be false alarms.  Considering
that the a measurement run should run for several hours, this
time-shifting for 30 seconds shouldn't really change the overall
results.

Signed-off-by: David Sommerseth <davids@redhat.com>