Introduce a new paths.py file to contain the installation paths.
This new paths.py file contains the paths to the configuration, the
rteval scripts, libs and data. It is used to know exactly where the
files are currently, depending on the installation parameters given
to setup.py. By default, this file contains paths allowing to use
rteval without having to install it.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>