cyclictest: Don't bother reporting 'variance'
The data used to do the mathematical reporting is based on histogram data,
where the calculated variance will be rather pointless to gather.

Imagine you have 3 samples in the historgram in the 2000µs slot, even if
you have quite a lot of data with the average measurements around 50µs,
the calculated "variance" for that time slot will be (1950^2)*3.  Even
though the final variance calculation will divide on number of samples
gathered, it will take an enormous amount of samples within a the lower
time slots to make this value interesting.

So rather, don't report or save the calculated variance. However, the
standard deviation is based on the variance, that value is far more sane
due to the needed square root function on the final variance value.  So
it gives a far more sane and interesting value.  The applied math in the
cyclictest module should be fine itself, it's just the variance value
that isn't much interesting on this kind of data source.

Signed-off-by: David Sommerseth <davids@redhat.com>