Added new measurement module, sysstat
This module runs in parallel with cyclictest and will every minute
use sysstat/sadc to collect all available system statistics.  These
statistics is both added as a single file to the tarball and embedded
into the summary.xml.

Also changed the polling time which checks if measurement and load modules
are alive from every second to every minute.

Signed-off-by: David Sommerseth <davids@redhat.com>