server/parser: BUG: The parsed values of mean and median was mixed up
The wrong XML tag was used in the database column for mean and median, causing
these two fields to be mixed up.  So the database saved the mean value in the median
column and the median value in the mean column.

For those who use this parser, the easiest approach is to just rename the columns in
database to correct this.  This should be done before installing this updated
parser definition.

Signed-off-by: David Sommerseth <davids@redhat.com>